Rename globalMessageIds.ts to messageIds.ts so that Lyra can find it
