+++ src/features/import/components/ImportDialog/Configure/Preview/FieldsPreview.tsx
@@ -1,4 +1,4 @@
-import globalMessageIds from 'core/i18n/globalMessageIds';
+import globalMessageIds from 'core/i18n/messageIds';
 import messageIds from 'features/import/l10n/messageIds';
 import PreviewGrid from './PreviewGrid';
 import useColumn from 'features/import/hooks/useColumn';
