+++ src/zui/ZUIAccessList/index.tsx
@@ -6,7 +6,7 @@ import { ZetkinObjectAccess } from 'core/api/types';
 import { ZetkinOfficial } from 'utils/types/zetkin';
 import ZUIRelativeTime from 'zui/ZUIRelativeTime';
 import { Msg, useMessages } from 'core/i18n';
-import globalMessageIds from 'core/i18n/globalMessageIds';
+import globalMessageIds from 'core/i18n/messageIds';
 import messageIds from 'zui/l10n/messageIds';
 
 interface ZUIAccessListProps {
